// src/Momentum.Frontend/apps/code-innovation-backlog/src/generated/services/Cycai_activitiesService.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

import type { Cycai_activitiesBase, Cycai_activities } from '../models/Cycai_activitiesModel';
import type { GetEntityMetadataOptions, EntityMetadata } from '@microsoft/power-apps/data/metadata/dataverse';
import type { IGetOptions, IGetAllOptions } from '../models/CommonModels';
import type { IOperationResult } from '@microsoft/power-apps/data';
import { dataSourcesInfo } from '../../../.power/schemas/appschemas/dataSourcesInfo';
import { getClient } from '@microsoft/power-apps/data';


export class Cycai_activitiesService {
  private static readonly dataSourceName = 'cycai_activities';

  private static readonly client = getClient(dataSourcesInfo);

  public static async create(record: Omit<Cycai_activitiesBase, 'cycai_activityid'>): Promise<IOperationResult<Cycai_activities>> {
    const result = await Cycai_activitiesService.client.createRecordAsync<Omit<Cycai_activitiesBase, 'cycai_activityid'>, Cycai_activities>(
      Cycai_activitiesService.dataSourceName,
      record
    );
    return result;
  }

  public static async update(id: string, changedFields: Partial<Omit<Cycai_activitiesBase, 'cycai_activityid'>>): Promise<IOperationResult<Cycai_activities>> {
    const result = await Cycai_activitiesService.client.updateRecordAsync<Partial<Omit<Cycai_activitiesBase, 'cycai_activityid'>>, Cycai_activities>(
      Cycai_activitiesService.dataSourceName,
      id,
      changedFields
    );
    return result;
  }

  public static async delete(id: string): Promise<void> {
    await Cycai_activitiesService.client.deleteRecordAsync(
      Cycai_activitiesService.dataSourceName,
      id);
  }

  public static async get(id: string, options?: IGetOptions): Promise<IOperationResult<Cycai_activities>> {
    const result = await Cycai_activitiesService.client.retrieveRecordAsync<Cycai_activities>(
      Cycai_activitiesService.dataSourceName,
      id,
      options
    );
    return result;
  }

  public static async getAll(options?: IGetAllOptions): Promise<IOperationResult<Cycai_activities[]>> {
    const result = await Cycai_activitiesService.client.retrieveMultipleRecordsAsync<Cycai_activities>(
      Cycai_activitiesService.dataSourceName,
      options
    );
    return result;
  }

  public static getMetadata(
    options: GetEntityMetadataOptions<Cycai_activities> = {}
  ): Promise<IOperationResult<Partial<EntityMetadata>>> {
    return Cycai_activitiesService.client.executeAsync({
      dataverseRequest: {
        action: "getEntityMetadata",
        parameters: {
          tableName: Cycai_activitiesService.dataSourceName,
          options: options as GetEntityMetadataOptions,
        },
      },
    });
  }
}